use poke instead of shake

+++ Lagrange.Core/Internal/Packets/Message/Element/Implementation/Extra/FriendShakeExtra.cs
@@ -5,7 +5,7 @@ namespace Lagrange.Core.Internal.Packets.Message.Element.Implementation.Extra;
 [ProtoContract]
 internal class FriendShakeExtra
 {
-    [ProtoMember(1)] public ushort FaceId { get; set; }
+    [ProtoMember(1)] public uint FaceId { get; set; }
 
-    [ProtoMember(7)] public ushort Strength { get; set; }
+    [ProtoMember(7)] public uint Strength { get; set; }
 }
